Broke ractive plugins out into separate files for shop & attendees.
