+++ brambling/templates/brambling/event/order/attendees.html
@@ -8,6 +8,7 @@
 	{{ block.super }}
 	<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/ractive/0.7.3/ractive.js"></script>
 
+	<script type="text/javascript" src="{% static "brambling/lib/ractive-events-tap.js" %}"></script>
 	<script type="text/javascript" src="{% static "brambling/attendees.ractive.js" %}"></script>
 
 	<script>
