+++ brambling/templates/brambling/event/order/register.html
@@ -22,10 +22,14 @@
 	{{ block.super }}
 
 	<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/ractive/0.7.3/ractive.js"></script>
+	<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/URI.js/1.16.1/URI.js"></script>
 
 	<script type="text/javascript" src="{% static "zenaida/js/bootstrap/carousel.js" %}"></script>
 	<script type="text/javascript" src="{% static "zenaida/js/bootstrap/modal.js" %}"></script>
 
+	<script type="text/javascript" src="{% static "brambling/lib/ractive-transitions-slide.js" %}"></script>
+	<script type="text/javascript" src="{% static "brambling/lib/ractive-transitions-fade.js" %}"></script>
+	<script type="text/javascript" src="{% static "brambling/lib/ractive-events-tap.js" %}"></script>
 	<script type="text/javascript" src="{% static "brambling/register.ractive.js" %}"></script>
 
 	<script type="text/javascript">
